Corrigindo tipo do campo PrecoUnidade

+++ ControleEstoque.Data/DTO/ItemPedidoDTO.cs
@@ -11,6 +11,6 @@ namespace ControleEstoque.Data.DTO
         public int Quantidade { get; set; }
 
         [JsonProperty("precoUnidade")]
-        public int PrecoUnidade { get; set; }
+        public double PrecoUnidade { get; set; }
     }
 }
